refactor: 봉사활동 카드 컴포넌트 분리
